Initial implementation of an OpenVDB based particle mesher modifier.
(NOTE) This is to be considered WIP as there are a few todos left. I'm uploading this here as a reference of my current work as discussed with Ton in IRC (easier than setting up a github repo).

(IMPORTANT) OpenVDB should be compiled and installed separately in order to test the patch. (Connect with me if you have any trouble.)

A crude documentation for the modifier can be found here: http://www.pasteall.org/56206/text. For now, I keep it as a branch (on my computer) of the blender manual.

Reviewers: campbellbarton, sergey, lukastoenne

Subscribers: bobizib, zeauro, SynaGl0w, lordodin, sergey, Lapineige

Differential Revision: https://developer.blender.org/D1008