Alembic: F-Curve cache modifier to lookup arbitrary property animations.
The main goal of this F-Curve modifier is to support cached animations
of camera focal length and is generalized to support looking up
animations of any single float property of the Alembic object pointed by
the modifier. Perhaps in the future we could also support array
properties, in which case the current scene frame would be the index in
the array.

The `Alembic Import` operator should automatically add a keyframe and F-
Curve cache modifier to imported cameras if they have an animated focal
length.

A cache is used to gather F-Curves modifier using cache files. This is
to make it easier to do this rather than iterating on all possible F-
Curves of a given object when updating a CacheFile data-block to figure
out whether or not it is using said CacheFile. Though it is not the
ideal solution, and should be handled by some depsgraph mechanism.

Reviewers: aligorith

Differential Revision: https://developer.blender.org/D2324